Started Missile Defense
Started to make the final project for Computer Science! Read Meggy Game
Proposal for the main idea for the game. I am making the blue dots at
the bottom (rockets). I pulled out the old Maze game for the cursor for
the game. Next time, I  will try to make the “rockets” launch towards
the cursor.